+++ xxl-job-admin/src/main/java/com/xxl/job/admin/core/model/XxlJobUser.java
@@ -7,17 +7,17 @@ import org.springframework.util.StringUtils;
  */
 public class XxlJobUser {
 	
-	private int id;
+	private long id;
 	private String username;		// 账号
 	private String password;		// 密码
 	private int role;				// 角色：0-普通用户、1-管理员
 	private String permission;	// 权限：执行器ID列表，多个逗号分割
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 
@@ -54,7 +54,7 @@ public class XxlJobUser {
 	}
 
 	// plugin
-	public boolean validPermission(int jobGroup){
+	public boolean validPermission(long jobGroup){
 		if (this.role == 1) {
 			return true;
 		} else {
