+++ xxl-job-core/src/main/java/com/xxl/job/core/biz/model/TriggerParam.java
@@ -8,7 +8,7 @@ import java.io.Serializable;
 public class TriggerParam implements Serializable{
     private static final long serialVersionUID = 42L;
 
-    private int jobId;
+    private long jobId;
 
     private String executorHandler;
     private String executorParams;
@@ -26,11 +26,11 @@ public class TriggerParam implements Serializable{
     private int broadcastTotal;
 
 
-    public int getJobId() {
+    public long getJobId() {
         return jobId;
     }
 
-    public void setJobId(int jobId) {
+    public void setJobId(long jobId) {
         this.jobId = jobId;
     }
